InMemory locking overhaul
No more maintain_lock() and refresh_lock().

Instead locks last forever, but are auto-dropped if the pid that locked
it is no longer alive.

You can still specify an unlock_after for use as a timer.

This means we no longer fork to maintain a refresher and no longer have
to wait until default timer runs out before getting a lock, but it does
mean we'll do a lot of ssh calls...

+++ modules/VRPipe/Persistent/InMemory.pm
@@ -22,7 +22,7 @@ Sendu Bala <sb10@sanger.ac.uk>.
 
 =head1 COPYRIGHT AND LICENSE
 
-Copyright (c) 2013 Genome Research Limited.
+Copyright (c) 2013-2015 Genome Research Limited.
 
 This file is part of VRPipe.
 
@@ -55,19 +55,6 @@ class VRPipe::Persistent::InMemory {
     use Time::HiRes qw(sleep);
     use Bytes::Random::Secure;
     use VRPipe::Interface::BackEnd;
-    
-    has '_maintenance_children' => (
-        is      => 'ro',
-        isa     => 'HashRef[ArrayRef[Int]]',
-        traits  => ['Hash'],
-        default => sub { {} },
-        handles => {
-            '_add_maintenance_child'    => 'set',
-            '_have_maintenance_child'   => 'exists',
-            '_delete_maintenance_child' => 'delete',
-            '_all_maintenance_children' => 'values'
-        },
-    );
     
     our $vrp_config   = VRPipe::Config->new();
     our $email_domain = $vrp_config->email_domain();
@@ -234,11 +221,14 @@ class VRPipe::Persistent::InMemory {
         }
     }
     
-    method lock (Str $key!, Int :$unlock_after = 900, Str :$key_prefix = 'lock', Bool :$non_exclusive = 0, Bool :$debug = 0) {
+    # default of unlock_after => 0 means the lock is held "forever", but any
+    # other process trying to get a lock will see if this process is still
+    # alive, and if not, the initial lock will be lost
+    method lock (Str $key!, Int :$unlock_after = 0, Str :$key_prefix = 'lock', Bool :$non_exclusive = 0, Bool :$debug = 0, Bool :$check_via_ssh = 1) {
         my $redis     = $self->_redis;
         my $redis_key = $key_prefix . '.' . $key;
         
-        my $valid_lock = $self->_check_existing_lock($redis_key, $debug);
+        my $valid_lock = $self->_check_existing_lock($redis_key, debug => $debug, check_via_ssh => $check_via_ssh);
         
         if ($non_exclusive) {
             # don't allow even a non_exclusive lock if we currently have an
@@ -250,7 +240,7 @@ class VRPipe::Persistent::InMemory {
         
         my $got_lock = $redis->set(
             $redis_key => $redis_key_value,
-            EX         => $unlock_after,
+            $unlock_after ? (EX => $unlock_after) : (),
             $non_exclusive ? () : ('NX')
         );
         
@@ -281,9 +271,10 @@ class VRPipe::Persistent::InMemory {
         return 0;
     }
     
-    method _check_existing_lock (Str $key!, Bool $debug = 0) {
+    method _check_existing_lock (Str $key!, Bool :$debug = 0, Bool :$check_via_ssh = 1) {
         my $redis = $self->_redis;
         my $val   = $redis->get($key);
+        
         if ($val) {
             my ($hostname, $pid) = split('!.!', $val);
             unless ($hostname && $pid) {
@@ -292,30 +283,37 @@ class VRPipe::Persistent::InMemory {
                 return;
             }
             
-            # check the current lock is for a currently living pid
-            if ($hostname eq hostname()) {
-                unless (kill(0, $pid)) {
-                    warn " lock for $key was initiated on host $hostname but the pid $pid is dead, so removing it\n" if $debug;
-                    $redis->del($key);
-                    return;
+            # check the current lock is for a currently living pid, or has a
+            # specified EX
+            my $ttl = $redis->ttl($key);
+            unless ($ttl && $ttl > 0) {
+                if ($hostname eq hostname()) {
+                    unless (kill(0, $pid)) {
+                        warn " lock for $key was initiated on host $hostname but the pid $pid is dead, so removing it\n" if $debug;
+                        $redis->del($key);
+                        return;
+                    }
                 }
-            }
-            else {
-                my $return;
-                eval {
-                    local $SIG{ALRM} = sub { die "alarm\n" };
-                    alarm 15;
-                    $return = $backend->ssh($hostname, "kill 0 $pid && echo pid_alive");
-                    alarm 0;
-                };
-                
-                if (!$return || $return !~ /pid_alive/) {
-                    warn " lock for $key was initiated on host $hostname but the pid $pid is dead, so removing it\n" if $debug;
-                    $redis->del($key);
-                    return;
+                elsif ($check_via_ssh) {
+                    my $return;
+                    eval {
+                        local $SIG{ALRM} = sub { die "alarm\n" };
+                        alarm 15;
+                        # we use perl for testing the pid because don't know
+                        # what shell the user has and what syntax to use
+                        $return = $backend->ssh($hostname, "perl -e 'print qq[pid_alive] if kill(0, $pid)'");
+                        alarm 0;
+                    };
+                    
+                    if (!$return || $return !~ /pid_alive/) {
+                        warn " lock for $key was initiated on host $hostname but the pid $pid is dead, so removing it\n" if $debug;
+                        $redis->del($key);
+                        return;
+                    }
                 }
             }
         }
+        
         return $val;
     }
     
@@ -323,47 +321,10 @@ class VRPipe::Persistent::InMemory {
         return $self->lock($key, unlock_after => $forget_after, key_prefix => 'note', non_exclusive => 1);
     }
     
-    method refresh_lock (Str $key!, Int :$unlock_after = 900, Str :$key_prefix = 'lock', :$lock_owners_pid?) {
-        my $redis_key = $key_prefix . '.' . $key;
-        return unless $self->_own_lock($redis_key, $lock_owners_pid || $$);
-        
-        my $redis = $self->_redis;
-        my $refreshed = $redis->expire($redis_key, $unlock_after);
-        unless ($refreshed) {
-            warn "pid $$ unable to refresh redis lock for $redis_key";
-            
-            # presumably the redis server went down and we lost the lock;
-            # if the server came back let's try and create the lock again
-            $refreshed = $redis->set($redis_key => 1, EX => $unlock_after, 'NX');
-        }
-        
-        return $refreshed;
-    }
-    
     method unlock (Str $key!, Str :$key_prefix = 'lock') {
         my $redis_key = $key_prefix . '.' . $key;
         return unless $self->_own_lock($redis_key);
-        
-        if ($self->_have_maintenance_child($redis_key)) {
-            my ($owner_pid, $child_pid) = @{ $self->_delete_maintenance_child($redis_key) };
-            if ($owner_pid == $$) {
-                kill(9, $child_pid);
-                waitpid $child_pid, 0;
-            }
-        }
-        
         return $self->_redis->del($redis_key);
-    }
-    
-    sub DEMOLISH {
-        my $self = shift;
-        foreach my $ref ($self->_all_maintenance_children) {
-            my ($owner_pid, $child_pid) = @$ref;
-            if ($owner_pid == $$) {
-                kill(9, $child_pid);
-                waitpid $child_pid, 0;
-            }
-        }
     }
     
     method forget_note (Str $key!) {
@@ -376,7 +337,9 @@ class VRPipe::Persistent::InMemory {
         if ($by_me) {
             return $self->_own_lock($redis_key);
         }
-        return $self->_redis->exists($redis_key);
+        #else
+        my $val = $self->_check_existing_lock($redis_key);
+        return defined $val ? 1 : 0;
     }
     
     method noted (Str $key!) {
@@ -389,12 +352,12 @@ class VRPipe::Persistent::InMemory {
         }
     }
     
-    method block_until_locked (Str $key!, Int :$check_every = 2, Str :$key_prefix = 'lock', Bool :$debug = 0, Int :$unlock_after?) {
+    method block_until_locked (Str $key!, Int :$check_every = 2, Str :$key_prefix = 'lock', Bool :$debug = 0, Int :$unlock_after = 0) {
         my $redis_key = $key_prefix . '.' . $key;
         return if $self->_own_lock($redis_key);
         my $sleep_time = 0.01;
-        $unlock_after ||= $deployment eq 'testing' ? 30 : 900;
-        while (!$self->lock($key, unlock_after => $unlock_after, key_prefix => $key_prefix, debug => $debug)) {
+        my $t          = time();
+        while (!$self->lock($key, unlock_after => $unlock_after, key_prefix => $key_prefix, debug => $debug, check_via_ssh => time() == $t)) {
             if ($sleep_time >= $check_every) {
                 $sleep_time = $check_every;
             }
@@ -404,60 +367,12 @@ class VRPipe::Persistent::InMemory {
             
             warn " - didn't get a lock on $redis_key, will try again in $sleep_time seconds\n" if $debug;
             sleep($sleep_time);
+            
+            # we'll check_via_ssh only ~once per minute
+            if (time() > $t + 60) {
+                $t = time();
+            }
         }
-    }
-    
-    method maintain_lock (Str $key!, Int :$refresh_every?, Int :$leeway_multiplier?, Str :$key_prefix = 'lock') {
-        my $redis_key = $key_prefix . '.' . $key;
-        $self->throw("maintain_lock() cannot be used unless we own the lock") unless $self->_own_lock($redis_key);
-        
-        unless ($self->_have_maintenance_child($redis_key)) {
-            unless ($refresh_every) {
-                $refresh_every = $deployment eq 'testing' ? 3 : 60;
-            }
-            
-            unless ($leeway_multiplier) {
-                if ($refresh_every <= 60) {
-                    $leeway_multiplier = 15;
-                }
-                elsif ($refresh_every <= 300) {
-                    $leeway_multiplier = 3;
-                }
-                else {
-                    $leeway_multiplier = 2;
-                }
-            }
-            my $survival_time = $refresh_every * $leeway_multiplier;
-            
-            # we used to just have an AnyEvent->timer that called refresh_lock
-            # periodically, but that timer doesn't fire while normal code is
-            # busy running. I also couldn't get a 'cede' type method working
-            # due to recursion in AnyEvent watchers. Instead we fork to
-            # periodically calls refresh_lock
-            # *** should we use Proc::FastSpawn or AnyEvent::Fork for greater
-            # speed and lower memory?
-            my $my_pid   = $$;
-            my $lock_pid = fork();
-            if (!defined $lock_pid) {
-                $self->throw("attempt to fork for lock failed: $!");
-            }
-            elsif ($lock_pid == 0) {
-                # child, initiate a lock that will end when the parent stops
-                # running
-                while (1) {
-                    kill(0, $my_pid) || last;
-                    last unless $self->_own_lock($redis_key, $my_pid);
-                    $self->refresh_lock($key, key_prefix => $key_prefix, unlock_after => $survival_time, lock_owners_pid => $my_pid);
-                    sleep $refresh_every;
-                }
-                
-                exit(0);
-            }
-            
-            $self->_add_maintenance_child($redis_key, [$$, $lock_pid]);
-        }
-        
-        return 1;
     }
     
     method enqueue (Str $key!, Str $value!) {
